update js, added x,y scale

+++ src/main/java/com/kurtlike/webserv/MathLab3/BadNewIO.java
@@ -91,7 +91,7 @@ public class BadNewIO implements GoodOldIO {
     }
 
     @Override
-    public double getNumberOfDivisions() {
+    public int getNumberOfDivisions() {
         return method.getNumberOfDivisions();
     }
