+++ src/main/java/com/kurtlike/webserv/MathLab3/methods/AbstractMethod.java
@@ -31,7 +31,7 @@ public abstract class AbstractMethod implements Method{
     }
 
     @Override
-    public double getNumberOfDivisions() {
+    public int getNumberOfDivisions() {
         return counter;
     }
 }
